requirements.txt
asgiref==3.5.1
Django==3.2.13
django-extensions==3.1.5
django-rest-framework==0.1.0
django-tenant-schemas==1.11.0
django-tenants==3.4.2
djangorestframework==3.13.1
ordered-set==4.1.0
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
